feat: реализация сервисов, миграций Liquibase, добавление связей Comment-Ad, изправления в user repository

+++ src/main/java/ru/skypro/homework/repository/UserRepository.java
@@ -1,5 +1,6 @@
 package ru.skypro.homework.repository;
 
+import jakarta.persistence.criteria.CriteriaBuilder;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.skypro.homework.model.user.User;
@@ -7,7 +8,7 @@ import ru.skypro.homework.model.user.User;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
+public interface UserRepository extends JpaRepository<User, Integer> {
 
     Optional<User> findByEmail(String username);
 }
